+++ CodeConverter/CSharp/VbMethodSyntaxExtensions.cs
@@ -1,25 +1,35 @@
 ﻿using System.Linq;
 using ICSharpCode.CodeConverter.Util;
 using Microsoft.CodeAnalysis;
+using VBasic = Microsoft.CodeAnalysis.VisualBasic;
+using VBSyntax = Microsoft.CodeAnalysis.VisualBasic.Syntax;
 
 namespace ICSharpCode.CodeConverter.CSharp
 {
     internal static class VbMethodSyntaxExtensions
     {
-        public static bool AllowsImplicitReturn(this Microsoft.CodeAnalysis.VisualBasic.Syntax.MethodBlockBaseSyntax node)
+        /// <summary>
+        /// Use in conjunction with <see cref="IMethodSymbolExtensions.ReturnsVoidOrAsyncTask(IMethodSymbol)" />
+        /// </summary>
+        public static bool MustReturn(this VBSyntax.MethodBlockBaseSyntax node)
         {
-            return !IsIterator(node) && node.IsKind(Microsoft.CodeAnalysis.VisualBasic.SyntaxKind.FunctionBlock, Microsoft.CodeAnalysis.VisualBasic.SyntaxKind.GetAccessorBlock);
+            return !IsIterator(node) && node.IsKind(VBasic.SyntaxKind.FunctionBlock, VBasic.SyntaxKind.GetAccessorBlock)
+                && !node.IsIterator();
         }
 
-        public static bool IsIterator(this Microsoft.CodeAnalysis.VisualBasic.Syntax.MethodBlockBaseSyntax node)
+        public static bool IsIterator(this VBSyntax.MethodBlockBaseSyntax node)
         {
-            var modifiableNode = node.IsKind(Microsoft.CodeAnalysis.VisualBasic.SyntaxKind.GetAccessorBlock) ? node.GetAncestor<Microsoft.CodeAnalysis.VisualBasic.Syntax.PropertyBlockSyntax>().PropertyStatement : node.BlockStatement;
-            return HasIteratorModifier(modifiableNode);
+            return GetMethodBlock(node).HasModifier(VBasic.SyntaxKind.IteratorKeyword);
         }
 
-        private static bool HasIteratorModifier(this Microsoft.CodeAnalysis.VisualBasic.Syntax.MethodBaseSyntax d)
+        public static bool HasModifier(this VBSyntax.MethodBaseSyntax d, VBasic.SyntaxKind modifierKind)
         {
-            return d.Modifiers.Any(m => SyntaxTokenExtensions.IsKind(m, Microsoft.CodeAnalysis.VisualBasic.SyntaxKind.IteratorKeyword));
+            return d.Modifiers.Any(m => SyntaxTokenExtensions.IsKind(m, modifierKind));
+        }
+
+        private static VBSyntax.MethodBaseSyntax GetMethodBlock(VBSyntax.MethodBlockBaseSyntax node)
+        {
+            return node.IsKind(VBasic.SyntaxKind.GetAccessorBlock) ? node.GetAncestor<VBSyntax.PropertyBlockSyntax>().PropertyStatement : node.BlockStatement;
         }
     }
 }