+++ CodeConverter/Util/IMethodSymbolExtensions.cs
@@ -1,4 +1,5 @@
 using System.Linq;
+using ICSharpCode.CodeConverter.Util.FromRoslyn;
 using Microsoft.CodeAnalysis;
 
 namespace ICSharpCode.CodeConverter.Util
@@ -13,6 +14,11 @@ namespace ICSharpCode.CodeConverter.Util
         public static (string Name, int TypeParameterCount, string ParameterTypes) GetUnqualifiedMethodSignature(this IMethodSymbol methodSymbol, bool caseSensitiveName)
         {
             return (caseSensitiveName ? methodSymbol.Name : methodSymbol.Name.ToLowerInvariant() , methodSymbol.TypeParameters.Length, GetParameterSignature(methodSymbol));
+        }
+
+        public static bool ReturnsVoidOrAsyncTask(this IMethodSymbol enclosingMethodInfo)
+        {
+            return enclosingMethodInfo.ReturnsVoid || enclosingMethodInfo.IsAsync && enclosingMethodInfo.ReturnType.GetArity() == 0;
         }
     }
 }